backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-oauth-toolkit==1.7.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Background tasks
Celery==5.3.4
redis==5.0.1

# Real-time
channels==4.2.0
channels-redis==4.1.0
daphne==4.0.0

# AI/ML
openai==1.3.7
langchain==0.0.340

# Document processing
python-docx==1.1.0
reportlab==4.0.7
pdfkit==1.0.0
WeasyPrint==60.2

# API integrations
requests==2.31.0
aiohttp==3.9.1

# Security
cryptography==41.0.7
django-ratelimit==4.1.0
python-jose[cryptography]==3.3.0

# Utilities
python-decouple==3.8
python-dateutil==2.8.2
pillow==10.1.0
whitenoise==6.6.0
gunicorn==21.2.0
uvicorn==0.24.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0